--- example-compilation/wasm-lib/.cargo/config.toml
@@ -1,2 +1,0 @@
-[build]
-target = "wasm32-unknown-unknown"
