--- example-compilation/wasm-lib/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "wasm-lib"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
