+++ example3/main.py
@@ -1,7 +1,9 @@
-from wasmtime import Store, Module, Instance # type: ignore
+from wasmtime import Store, Module, Instance, Func, FuncType, ValType # type: ignore
 store = Store()
 module = Module.from_file(store.engine, 'sum.wat')
-instance = Instance(store, module, [])
+print_result = Func(store, FuncType([ValType.i32()], []), lambda x: print(f"Result found: {x}"))
+
+instance = Instance(store, module, [print_result])
 sum = instance.exports(store)["sum"]
-result = sum(store, 11, 92)
+result = sum(store, 20, 22)
 print(result)
