+++ example3/main.rs
@@ -5,8 +5,13 @@ fn main() -> Result<()> {
     let module = Module::from_file(&engine, "sum.wasm")?;
 
     let mut store = Store::new(&engine, ());
+    let print_result = Func::wrap(&mut store, |x: i32| {
+        println!("Result found: {}", x);
+    });
 
-    let instance = Instance::new(&mut store, &module, &[])?;
+    let imports = [print_result.into()];
+
+    let instance = Instance::new(&mut store, &module, &imports)?;
     let run = instance.get_typed_func::<(i32, i32), i32>(&mut store, "sum")?;
     let result= run.call(&mut store, (20, 22))?;
     println!("{result}");
