--- example4/.gitignore
@@ -1,3 +1,0 @@
-target/
-Cargo.lock
-hello.wasm
